перенос команд call_data в отдельный файл
